+++ tests/Fixtures/Mitoop/LaravelEfficientFormRequest/EfficientSceneFormRequest.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Mitoop\LaravelEfficientFormRequest;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+abstract class EfficientSceneFormRequest extends FormRequest {}