+++ tests/Fixtures/Mitoop/LaravelQueryBuilder/AbstractFilter.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Mitoop\LaravelQueryBuilder;
+
+abstract class AbstractFilter
+{
+    public function rules(): array
+    {
+        return [];
+    }
+}