+++ tests/Fixtures/Models/User.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Admin9\ScrambleExtensions\Tests\Fixtures\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class User extends Model
+{
+    protected $guarded = [];
+}